Compiling Expressions
A bunch of things that were global variables in the chapter have been
shoved into the Parser struct here: the global scanner in scanner.c is a
straight-up field here, while compilingChunk is a pointer field to a
chunk externally supplied to the compile function.  We'll see how this
holds up going forward.