Hash Tables
Like in the Strings chapter, the compiler needs to work with the VM's
strings table.  Unlike the objects list, the VM's strings table doesn't
move around in memory, so the compiler just takes a pointer to it and
passes it around in its internal parser.

Care is taken in compiler_test.c and value_test.c to free the strings
table before the objects list, since the former holds pointers to the
latter and the latter frees its own contents.