deduplicate constants and variable names during parsing
When the compiler needs to emit a value index for constants and variable
names, do a linear search to see if it already exists and return it if
so.  This means the compiler does more work, but it raises the value
limit from 256 values/variable names to 256 *unique* values/variable
names, which is probably worthwhile.

This change assumes that all values in a chunk's values array are
immutable, otherwise this deduplication isn't safe.