Local Variables
Compared to the chapter, the 'current' variable representing the
compiler state in compiler.c is instead a field named 'currentCompiler'
in the Parser pseudo-global.  It can't be a bona-fide global due to the
unit tests, and it can't be called 'current' because the Parser
pseudo-global already has a Token field with that name.