add local/global 'const' declarations, fix compile error memory access
Local constants are easy, since locals must be declared before use;
adding a 'constant' field lets the compiler easily track if a constant
is assigned to.

Globals are trickier, since there's no pre-existing tracking for them.
Worse, they can be assigned to *before* they're declared, since they're
late-binding, but the compiler is single-pass only.  A globals table is
added to the Parser pseudo-global, with a 'true' set when a const global
is defined, and 'false' set when a global is declared or assigned to.
If at any of these points an entry already exists and conflicts with
what's happening, an appropriate error is raised.

Separately to all of this, there's invalid memory access after compiling
fails with variables declared in vm.c and more interpretation with
variables is attempted; objects were being freed too early.  The fix is
to always extend the objects chain when compiling.  This needs to be
ported to mainline too.