add 'continue' statement
A 'continue' statement jumps to the start of the nearest enclosing loop.
A 'continue' outside of a loop is compile error.  Local variables on the
stack are correctly handled.

Loop starts and their scope depth are tracked by the compiler; this
information is used to set the loop back-jump target and pop the correct
number of local variables from the stack when a 'continue' token is
encountered.