initialize currentCompiler field of Parser in compiler.c
Back in the Local Variables chapter, the compiler is created as a
global: "Compiler* current = NULL;" but I made it into a
"currentCompiler" field as part of the Parser struct instead.  I forgot
to initialize it to NULL, so this fixes that.

I never noticed because its starting value hasn't been used for anything
yet, so it's impossible to write a test for.