Closures
This is the hardest chapter of the book so far.  The code-first approach
the book takes here makes it hard to see the big picture.

Closures in jlox were just maps of local variable names to values that
themselves were kept by JVM closures while inner functions needed them.
Closures in clox are very different: they instead associate functions
that use local variables from enclosing scopes with an array of
*upvalues*.

An upvalue is like a mini-closure for a single local variable, serving
as a reference to that variable as well as a sort of "life boat" for it
when its stack slot "sinks".  There is exactly one upvalue for every
local variable used by any number of enclosing functions.

The compiler detects and tracks the use of local variables in enclosing
functions.  When such a function is finalized, the compiler emits
'OP_CLOSURE' followed by the upvalues that the VM should bundle with the
function in a new closure; preparing the life boats, so to speak.

The VM keeps a sorted linked list of upvalues whose local variables are
still on the value stack.  If the enclosure is a scope,
'OP_CLOSE_UPVALUE' "rescues" (closes) the local variable into its
upvalue; the upvalue fixes its pointer to point into itself so access
via 'OP_GET_UPVALUE' and 'OP_SET_UPVALUE' work without changes.  If the
enclosure is a function, 'OP_RETURN' automatically rescues all local
variables into upvalues that would otherwise be popped.

Closure upvalue arrays track upvalues transitively: every function
between a local variable and its use in an enclosing function keeps a
tab on that variable's upvalue.  These arrays *must* hold pointers to
upvalues in order to facilitate sharing that would otherwise be
impossible.

In terms of memory lifetime, upvalues are allocated objects that
initially appear in the VM's sorted upvalue linked list and are shared
with closures as the VM encounters functions.  When a local variable is
rescued into an upvalue, the VM removes it from its linked list, but the
functions hold pointers to them for as long as they need to.  Without
garbage collection, the upvalue remains allocated even after all
referring closures are inaccessible, so they're only freed when the VM
frees everything at the very end of execution.