Garbage Collection
There's lots of changes here, but this is a fairly easy chapter compared
to the one for Closures.  Differences from the chapter's code:

- Due to avoiding globals, the objects chain and all of the data fields
  related to garbage collection go into a dedicated GC struct, instead
  of the VM struct like in the book.  More details below.

- The code for marking roots and fixing weak references is moved closer
  to the VM and compiler and is set via 'hooks' in the GC struct; also
  see below.

- Most stack-only temporary objects are protected from early sweeping
  slightly differently; also see below.

- The DEBUG_LOG_GC and DEBUG_STRESS_GC defines are replaced with runtime
  flags and are disabled by default; send "--log-gc" and "--stress-gc"
  to the main interpreter to enable them.

The GC struct holds all the data for garbage collection that the chapter
throws into its VM global: allocated byte count, the auto-GC threshold
and the 'gray' stack.  The GC struct also takes full ownership of the
objects chain from the VM; pretty much all places that allocate and free
objects need a pointer to a GC now instead of an objects chain.

The VM gets its own GC struct in exchange for that objects chain.  Test
code can create standalone GCs when a VM isn't present.

The GC struct has 'markRoots' and 'fixWeak' hooks (function pointers)
for marking GC roots and fixing weak references.  The VM and compiler
have different roots and set different custom 'markRoots' hooks.  The VM
lends its GC struct to the compiler, which wraps any existing hooks
while it runs, and restores them when its done.

Stack-only temporary objects are guarded from sweeping by pushing and
popping them from the VM's stack in the chapter.  Here, the GC struct
gets its own 'temp' stack that serves the same purpose without needing a
VM.  Objects are added and removed using the 'pushTemp' and 'popTemp'
functions.

GC bugs can be easily tracked down thanks to AddressSanitizer
(-fsanitize=address): simply stress the GC (done automatically in tests)
and ASAN will print stack traces for the locations of the segfault, the
memory free and allocation.

Despite the differences, as much effort as possible has been made to
keep the structure of the code here close to what's in the book, which
is needed for upcoming chapters.