unify string allocation; find interned concat strings before allocation
All strings are now allocated and initialized by the new concatString
function; copyString calls it with an empty second string and vm.c uses
it to concatenate strings instead of allocating memory itself and
calling takeString, which is now gone.  This is a prerequisite for small
string optimization for ObjString.

concatString does the following:

1. Calculate the combined hash of the input strings.
2. Check if the combined input is interned and just return it if so.
3. Allocate char storage for a new unique string.
4. Initialize a new ObjString.
5. Intern the new ObjString.

hashString is now hashAnotherString and continues hashing a string
starting from an existing hash value; concatString uses it to get the
hash of the combined input strings starting with the hash of the first
string and processing the second string.

tableFindString is now tableJoinedStringsEntry, and either finds an
interned copy of two yet-to-be-joined strings, or an appropriate Entry
to insert them into once joined; concatString uses this Entry to either
return an already-interned string or intern a unique string
concatenation.

allocateString is stripped down to just allocating a blank ObjString;
concatString now initializes and interns strings instead.  Unlike
before, concatStrings reuses the first interned string hash Entry result
to intern its new ObjString using the new tableSetEntry function, which
is just like tableSet, except it inserts into an existing Entry.

Benchmark results (compiled with -O3, means of 5 batches of 10 runs each
except for table_bench):

- clock_bench: before=634.137ms after=633.658ms (0.1% faster)
- fib_bench: before=526.202ms after=563.905ms (7.2% slower)
- loop_math_bench: before=618.297ms after=546.784ms (11.6% faster)
- string_bench: before=45.969ms after=37.7602ms (17.9% faster)
- table_bench: before=585.773us after=558.292us (4.7% faster)
- zoo_bench: before=520.476ms after=500.824ms (3.8% faster)

These results are hard to understand: nothing here is supposed to change
much apart from string_bench.  Maybe object code is being rearranged
internally to cause these changes?