small string optimization for ObjString
When a new unique string is created, store its characters in the
ObjString directly instead of allocating separate memory if it fits.

Benchmark results (compiled with -O3, means of 5 batches of 10 runs each
except for table_bench):

- clock_bench: before=633.658ms after=634.115ms (0.1% slower)
- fib_bench: before=563.905ms after=597.096ms (5.9% slower)
- loop_math_bench: before=546.784ms after=563.944ms (3.1% slower)
- string_bench: before=37.7602ms after=39.0588ms (3.4% slower)
- table_bench: before=558.292us after=544.896us (2.4% faster)
- zoo_bench: before=500.824ms after=494.286ms (1.3% faster)

The numbers of the longer benchmarks tend to bias towards 'bands', e.g.
fib_bench runs 2 batches in around 650ms, then 3 in 560ms.  The drop in
numbers might just come down to bad luck with batches.

The lack of performance improvement for string_bench is surprising; this
change should have halved the number of memory allocations for unique
strings, and yet there's no significant improvement at all.  Maybe it's
bottlenecked on something else?