Makefile: use -Og for debug/coverage modes; link with -O3 for release
GCC turns off some optimizations that get in the way of debugging when
-Og is used versus -O1, even though clang treats them identically.

The -O flag is supposed to be included when linking with -flto, so now
it's enabled.  Release mode benchmarks before and after (means of 5
batches of 10 runs each except for table_bench):

- clock_bench: before=637.399ms after=636.886ms (0.1% faster)
- fib_bench: before=475.379ms after=475.328ms (0.1% faster)
- loop_math_bench: before=524.741ms after=513.918ms (2.1% faster)
- string_bench: before=35.829ms after=36.548ms (2.0% slower)
- table_bench: before=455.795us after=463.73us (1.7% slower)
- zoo_bench: before=390.047ms after=390.406ms (0.1% slower)

Most of these look like random fluctuation; loop_math_bench suffers from
multi-modal "clustered" results, e.g. here are the before batches:

1. mean 525.611ms, confidence interval +- 0.426149%
2. mean 559.047ms, confidence interval +- 0.654428%
3. mean 526.665ms, confidence interval +- 0.532489%
4. mean 525.271ms, confidence interval +- 0.348778%
5. mean 487.112ms, confidence interval +- 0.306148%

... and here's are the after batches:

1. mean 489.059ms, confidence interval +- 0.557071%
2. mean 530.423ms, confidence interval +- 2.343526%
3. mean 530.582ms, confidence interval +- 1.545929%
4. mean 493.778ms, confidence interval +- 1.517541%
5. mean 525.747ms, confidence interval +- 0.915496%

The means of the means here don't tell the whole story; something to
keep in mind going forward.