add string methods: size, substr and parsenum
Strings now have a few methods:

- size() - return the length of the string in bytes (no null terminator)
- substr(s, e) - return the substring from positions s to e (exclusive)
- parsenum() - parse a number from the string and return it

If substr gets a negative s or e value, the string length plus one is
added to the value.  s and e are clamped between zero and the string
length.  After all of this, if s is greater than or equal to e then an
empty string is returned, otherwise the substring starting at s and up
to but not including e is returned.

parsenum ignores leading and trailing whitespace.  parsenum returns nil
if a number cannot be parsed from the string, or if extra characters are
found beyond the pading whitespace and the number itself.