add native function: chr
chr(n) converts the byte value n into a single character string whose
first byte has the value n, i.e. `chr(n)[0] == n` should be true.

The range of valid values for chr depends on the range of the C 'char'
type for the target platform; for this computer it's -128 to 127
inclusive.