add native function: exit
exit(n) terminates the program with n as the numeric exit code.

The exit code must be between zero and the maximum value of the C
'unsigned char' type of the platform (typically 255).

The "InterpretMulti.ForceGC" test in interpret_test.c needs to be
updated to continue covering the code branches it's designed to test.